backend/requirements.txt
# Flask core
Flask==3.0.0
Flask-CORS==4.0.0

# Supabase
supabase==2.27.2

# HTTP requests
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
